add interface and move tests
