--- dev-requirements.txt
@@ -1,1 +1,0 @@
-Django==3.2.3
