+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from .unittests import *
+from .integrationtests import *
