--- tests/pytest.ini
@@ -1,8 +1,0 @@
-# pytest.ini
-[pytest]
-minversion = 6.0
-addopts = -ra -q
-testpaths =
-    command
-    services
-    .
