+++ tests/unittests/commands/in/data_migrations/0001_first.py
@@ -4,13 +4,13 @@ Generated by data_migration 0.0.1a
 File: 0001_first.py
 Created: 2021-05-22 11:23
 """
-from tests.commands.test_migrate import set_some_value
+from tests.unittests.commands.test_migrate import set_some_value
 
 
 class Node:
     name = '0001_first'
     dependencies = ()
-    migration_dependencies = ('test_app.0001_first', )
+    migration_dependencies = ()
     routines = [
         set_some_value
     ]
