+++ tests/unittests/commands/test_aliases.py
@@ -1,8 +1,8 @@
-from unittest import mock, skipIf
+from unittest import mock
 
 from django.core.management import call_command
 from django.test import TestCase
-from tests.test_app.helper import ResetDirectoryContext
+from tests.unittests.test_app.helper import ResetDirectoryContext
 
 
 class AliasCommandTestCase(TestCase):
@@ -18,3 +18,9 @@ class AliasCommandTestCase(TestCase):
             migrate_command.return_value = ''
             call_command('django_makemigrations', ['test_app'])
             migrate_command.assert_called_once()
+
+    @mock.patch('django.core.management.commands.squashmigrations.Command.handle')
+    def test_django_squashmigraitons(self, migrate_command):
+        migrate_command.return_value = ''
+        call_command('django_squashmigrations', 'test_app', '0001_initial')
+        migrate_command.assert_called_once()
