+++ tests/unittests/services/data_migrations/0002_auto.py
@@ -4,7 +4,7 @@ Generated by data_migration 0.0.1a
 File: 0002_auto.py
 Created: 2021-05-22 11:23
 """
-from tests.services.test_graph import set_some_value
+from tests.unittests.services.test_graph import set_some_value
 
 
 class Node:
