+++ tests/unittests/test_app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TestAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tests.test_app'
+    name = 'tests.unittests.test_app'
